refactor!: Move destinations to `cmd/destinations.go`
